v3.0

+++ src/main/java/org/example/DTO/DeleteTaskRequest.java
@@ -10,7 +10,7 @@ import org.springframework.stereotype.Component;
 @Scope("prototype")
 public class DeleteTaskRequest {
     @Positive(message = "id не может быть <=0")
-    private int user_id;
+    private Long user_id;
     @Positive(message = "id не может быть <=0")
-    private int task_id;
+    private Long task_id;
 }
